mesh: Make bounds computation reusable.
