+++ VisualPinball.Unity/VisualPinball.Unity/VPT/MeshComponent.cs
@@ -15,7 +15,10 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 using System;
+using System.Collections.Generic;
+using Unity.Mathematics;
 using UnityEngine;
+using VisualPinball.Engine.Math;
 using VisualPinball.Engine.VPT;
 using VisualPinball.Engine.VPT.Table;
 using Mesh = VisualPinball.Engine.VPT.Mesh;
@@ -96,6 +99,22 @@ namespace VisualPinball.Unity
 				var mr = gameObject.AddComponent<MeshRenderer>();
 				mr.sharedMaterial = material.ToUnityMaterial(matProvider, texProvider);
 			}
+		}
+
+		protected static Bounds CalculateBounds(IEnumerable<DragPointData> dragPoints, float margin = 0, float sizeZ = 0)
+		{
+			var min = new float3(float.MaxValue, float.MaxValue, float.MaxValue);
+			var max = new float3(float.MinValue, float.MinValue, float.MinValue);
+			foreach (var t in dragPoints) {
+				var p = (float3)t.Center.ToUnityVector3();
+				min = math.min(min, p);
+				max = math.max(max, p);
+			}
+			var middle = min + (max - min) / 2;
+			var sizeXy = max - min;
+			var size = new float3(sizeXy.x, sizeXy.y, sizeZ > 0 ? sizeZ : sizeXy.z) + margin * new float3(1f, 1f, 1f);
+
+			return new Bounds(middle, size);
 		}
 
 		private void UpdateMesh()
